requirements.txt
PyQt5>=5.15.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pytesseract>=0.3.10
googletrans==4.0.0-rc1